Namespace error classes in Slack::Web::Api::Errors
I moved `Slack::Web::Api::Error` to
`Slack::Web::Api::Errors::SlackError` but left the orginal constant
aliased to the new one.
