+++ lib/slack-ruby-client.rb
@@ -18,8 +18,9 @@ rescue LoadError
   # ignore, only used in users_search
 end
 require_relative 'slack/web/config'
+require_relative 'slack/web/api/errors/slack_error'
+require_relative 'slack/web/api/errors/too_many_requests_error'
 require_relative 'slack/web/api/error'
-require_relative 'slack/web/api/too_many_requests_error'
 require_relative 'slack/web/faraday/response/raise_error'
 require_relative 'slack/web/faraday/connection'
 require_relative 'slack/web/faraday/request'
