+++ lib/slack/web/api/error.rb
@@ -1,14 +1,7 @@
 module Slack
   module Web
     module Api
-      class Error < ::Faraday::Error
-        attr_reader :response
-
-        def initialize(message, response = nil)
-          @response = response
-          super message
-        end
-      end
+      Error = Errors::SlackError
     end
   end
 end
