--- lib/slack/web/api/too_many_requests_error.rb
@@ -1,21 +1,0 @@
-module Slack
-  module Web
-    module Api
-      class TooManyRequestsError < ::Faraday::Error
-        attr_reader :response
-
-        def initialize(response)
-          @response = response
-        end
-
-        def message
-          "Retry after #{retry_after} seconds"
-        end
-
-        def retry_after
-          response.headers['retry-after'].to_i
-        end
-      end
-    end
-  end
-end
