+++ lib/slack/web/faraday/response/raise_error.rb
@@ -5,11 +5,11 @@ module Slack
         class RaiseError < ::Faraday::Response::Middleware
           def on_complete(env)
             if env.status == 429
-              fail Slack::Web::Api::TooManyRequestsError, env.response
+              fail Slack::Web::Api::Errors::TooManyRequestsError, env.response
             elsif (body = env.body) && body['ok']
               return
             else
-              fail Slack::Web::Api::Error.new(body['error'], env.response)
+              fail Slack::Web::Api::Errors::SlackError.new(body['error'], env.response)
             end
           end
         end
