+++ spec/slack/web/api/endpoints/custom_specs/auth_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Slack::Web::Api::Endpoints::Auth do
   let(:client) { Slack::Web::Client.new }
   context 'without auth', vcr: { cassette_name: 'web/auth_test_error' } do
     it 'fails with an exception' do
-      expect { client.auth_test }.to raise_error Slack::Web::Api::Error, 'not_authed'
+      expect { client.auth_test }.to raise_error Slack::Web::Api::Errors::SlackError, 'not_authed'
     end
   end
   context 'with auth', vcr: { cassette_name: 'web/auth_test_success' } do
@@ -16,7 +16,7 @@ RSpec.describe Slack::Web::Api::Endpoints::Auth do
     it 'fails with an specific exception' do
       begin
         client.auth_test
-      rescue Slack::Web::Api::TooManyRequestsError => e
+      rescue Slack::Web::Api::Errors::TooManyRequestsError => e
         expect(e.message).to eq('Retry after 3600 seconds')
         expect(e.retry_after).to eq(3600)
       end
