+++ spec/slack/web/api/mixins/channels_spec.rb
@@ -25,7 +25,7 @@ RSpec.describe Slack::Web::Api::Mixins::Channels do
       expect(subject.channels_id(channel: '#general')).to eq('ok' => true, 'channel' => { 'id' => 'CDEADBEEF' })
     end
     it 'fails with an exception' do
-      expect { subject.channels_id(channel: '#invalid') }.to raise_error Slack::Web::Api::Error, 'channel_not_found'
+      expect { subject.channels_id(channel: '#invalid') }.to raise_error Slack::Web::Api::Errors::SlackError, 'channel_not_found'
     end
   end
 end
